Allow Metro to run directly from source
