+++ packages/metro/package.json
@@ -70,6 +70,7 @@
     "babel-jest": "^26.6.3",
     "dedent": "^0.7.0",
     "jest-snapshot": "^26.5.2",
+    "metro-babel-register": "0.71.0",
     "metro-memory-fs": "0.71.0",
     "metro-react-native-babel-preset": "0.71.0",
     "metro-react-native-babel-transformer": "0.71.0",
